Move history.py TreeLoop conversion from recorder.rs to history.rs
recorder.rs was incorrectly attributed to history.py's History class,
but its actual upstream counterpart is opencoder.py's Trace (raw op
storage). The TreeLoop construction logic and test_history.py parity
tests belong in history.rs alongside the TreeLoop definition.

- Add `Trace::into_parts()` to expose raw recorder data
- Add `impl From<recorder::Trace> for TreeLoop` in history.rs
- Replace incorrect `pub type History = TreeLoop` with the From impl
- Move 5 history.py parity tests from recorder.rs to history.rs
- Update recorder.rs docstring to reference opencoder.py
- Simplify `TraceCtx::into_tree_loop` to use `into_parts()` directly
